Adding the initial pass at documentation;
I'm aiming for having the Style Guide more exposed in the documentation.
I'm expecting to be able to reference it if people start asking for
things that would go against the guide. Also if anything I write that
follows the guide but isn't accessible then maybe I can contribute back
to the guide.