Refactor the theme selection;
So the idea here is that the themes are only different based on the
variables for the theme files. So this will change the theme variables
to have `local` prefixes to keep the namespace regular. I decided to
keep this like this until I get more feedback, because I think I could
also just manually add the local to everything in the old `tmuxtheme`
files. I chose this though because I really like how the
`catppuccin-selected-theme` can be used for local debugging or testing
out new colors if things were ever to need to be changed. If what that
`sed` and `source` command is doing is unclear, I can do better about
commenting around it.